rename helpers.js to onResize.js inside utils folder
